+++ src/Sql.Cmdlets/Get-Scalar.cs
@@ -46,6 +46,12 @@ public class GetScalarCommand: PSCmdlet {
 	public int Timeout { get; set; } = 30;
 
 	/// <summary>
+	/// The transaction to use, if any.
+	/// </summary>
+	[Parameter]
+	public IDbTransaction? Transaction { get; set; }
+
+	/// <summary>
 	/// Performs execution of this command.
 	/// </summary>
 	protected override void ProcessRecord() {
@@ -53,6 +59,6 @@ public class GetScalarCommand: PSCmdlet {
 			? PositionalParameters.ToOrderedDictionary()
 			: Parameters.Cast<DictionaryEntry>().ToDictionary(entry => entry.Key.ToString()!, entry => entry.Value);
 
-		WriteObject(Connection.ExecuteScalar(Command, parameters, new(Timeout: Timeout, Type: CommandType)));
+		WriteObject(Connection.ExecuteScalar(Command, parameters, new(Timeout, Transaction, CommandType)));
 	}
 }